requirements.txt
AllPairs==2.0.1
cffi==1.1.2
cryptography==0.9.1
decorator==3.4.2
demjson==2.2.2
enum34==1.0.4
fake-factory==0.5.2
functools32==3.2.3.post1
idna==2.0
ipaddress==1.0.7
jsonpath==0.54
jsonpath-rw==1.4.0
jsonschema==2.5.1
MySQL-python==1.2.5
ndg-httpsclient==0.4.0
ply==3.6
pyasn1==0.1.8
pycparser==2.14
pyOpenSSL==0.15.1
PyYAML==3.11
requests==2.7.0
robotframework>=2.8.7
simplejson==3.7.3
six==1.9.0
wheel==0.24.0
